Модель User вынесена в папку Models
